minor reorganization of tests
